feat: 实现敌人智能射击与子弹方向计算
本次提交实现了以下功能：
1. 敌人可智能向玩家方向射击，包含射击冷却机制
2. 子弹移动方向改为浮点数计算，支持任意角度发射
3. 子弹渲染时根据移动方向进行旋转
4. 优化了子弹生成位置计算，确保居中对齐
5. 敌人和玩家子弹均根据移动方向设置正确轨迹
6. 添加了必要的辅助方法和数据结构支持新功能

+++ src/bullet.cpp
@@ -19,7 +19,6 @@ s32 Bullet::init(const std::string& img_texture_path)
         else if (spawner()->IsEnemy())
         {
             // 敌人发射的子弹，向下发射
-            // TODO 不是严格的向下，需要计算角度
             set_move_dir_x(0);
             set_move_dir_y(1);
         }
@@ -50,8 +49,8 @@ s32 Bullet::UpdatePosition(s64 now_ms)
         dt = 0.0f;
     }
 
-    position_.x += static_cast<f32>(move_dir_x()) * speed() * dt;
-    position_.y += static_cast<f32>(move_dir_y()) * speed() * dt;
+    position_.x += move_dir_x() * speed() * dt;
+    position_.y += move_dir_y() * speed() * dt;
 
     // 边界检测
     auto this_rect = GetRect();
